3-lab-controlAccessAPI/source/cdk/requirements.txt
aws-cdk.core==1.96.0
aws-cdk.aws-iam==1.96.0 
aws-cdk.aws-lambda==1.96.0
aws-cdk.aws-apigateway==1.96.0
